+++ CHANGELOG.md
@@ -15,6 +15,10 @@ You can find its changes [documented below](#021-2024-12-27).
 
 This release has an [MSRV][] of 1.82.
 
+### Fixed
+
+* Colors in `XyzD65` are serialized as `xyz-d65` rather than `xyz`. ([#118][] by [@waywardmonkeys][])
+
 ## [0.2.1][] (2024-12-27)
 
 This release has an [MSRV][] of 1.82.
@@ -95,6 +99,7 @@ This is the initial release.
 [#109]: https://github.com/linebender/color/pull/109
 [#111]: https://github.com/linebender/color/pull/111
 [#113]: https://github.com/linebender/color/pull/113
+[#118]: https://github.com/linebender/color/pull/118
 
 [Unreleased]: https://github.com/linebender/color/compare/v0.2.1...HEAD
 [0.2.1]: https://github.com/linebender/color/releases/tag/v0.2.1